refactor(server): split ws handler into message and typing services
- extract send/fanout/delivery into ws/message_service.rs
- extract typing, read receipts, presence, caches into ws/typing_service.rs
- handler.rs slimmed to socket lifecycle + dispatcher + voice/canvas
- update routes/groups.rs and routes/messages.rs imports
