+++ apps/server/src/ws/mod.rs
@@ -1,2 +1,4 @@
 pub mod handler;
 pub mod hub;
+pub mod message_service;
+pub mod typing_service;
